clean up application insights setup code
